+++ .devcontainer/devcontainer.json
@@ -34,7 +34,5 @@
   // "remoteUser": "root"
 
   // https://www.reddit.com/r/github/comments/1kdqbpk/comment/mqo4w9i/
-  "mounts":[
-    "source=/home/youboon2/.gnupg,target=/home/node/.gnupg,type=bind,consistency=cached"
-  ]
+  "mounts": ["source=/home/youboon2/.gnupg,target=/home/node/.gnupg,type=bind,consistency=cached"]
 }
