+++ docs/content/00-general/01-database/index.md
@@ -76,12 +76,18 @@ This table stores information about LTI Tool Providers connected to this system.
 * `id` - autogenerated database ID field
 * `name` - user-friendly name of the LTI Tool Provider
 * `lti13` - boolean value representing if the LTI Tool Provider supports LTI 1.3 connections.
-  * **LTI 1.3 is currently not implemented for LTI Tool Providers**
 * `key` - {{% badge %}}Both{{% /badge %}} shared key used to identify the LTI Tool Provider
 * `launch_url` - {{% badge %}}LTI 1.0{{% /badge %}} the launch URL for LTI 1.0 connections to the LTI Tool Provider
-* `domain` - {{% badge %}}LTI 1.0{{% /badge %}} the base domain name of the LTI Tool Provider
-* `custom` - {{% badge %}}LTI 1.0{{% /badge %}} custom launch variables provided as part of the LTI 1.0 connection to the LTI Tool Provider
+* `domain` - {{% badge %}}Both{{% /badge %}} the base domain name of the LTI Tool Provider
+* `custom` - {{% badge %}}Both{{% /badge %}} custom launch variables provided as part of the LTI 1.0 connection to the LTI Tool Provider
 * `use_section` - {{% badge %}}Both{{% /badge %}} flag for some LTI Tool Providers that require special handling. This is not implemented within this library but may be handy for other uses.
+* `client_id` - {{% badge %}}LTI 1.3{{% /badge %}} client ID provided to the LTI provider
+* `deployment_id` - {{% badge %}}LTI 1.3{{% /badge %}} deployment ID provided to the LTI provider
+* `keyset_url` - {{% badge %}}LTI 1.3{{% /badge %}} JWKS keyset URL for the LTI Provider 
+* `auth_url` - {{% badge %}}LTI 1.3{{% /badge %}} Authentication URL for the LTI Provider
+* `redirect_urls` - {{% badge %}}LTI 1.3{{% /badge %}} Possible redirect URLs for the LTI Provider
+* `scopes` - {{% badge %}}LTI 1.3{{% /badge %}} Available scopes for the LTI provider (unused)
+* `claims` - {{% badge %}}LTI 1.3{{% /badge %}} Available claims for the LTI provider (unused)
 * `createdAt` - {{% badge %}}Auto{{% /badge %}} creation timestamp
 * `updatedAt` - {{% badge %}}Auto{{% /badge %}} last updated timestamp
 
@@ -91,3 +97,24 @@ This table stores the secrets associated with an LTI Tool Provider. It is stored
 
 * `key` - {{% badge %}}Auto{{% /badge %}} shared key used to identify the LTI Tool Provider
 * `secret` - {{% badge %}}LTI 1.0{{% /badge %}} shared secret with the LTI Tool Provider
+* `public` - {{% badge %}}LTI 1.3{{% /badge %}} public key used in connections with the LTI Provider
+* `private` - {{% badge %}}LTI 1.3{{% /badge %}} paired private key used in connections with the LTI Provider
+
+### lti_provider_logins
+
+This table keeps track of outgoing LTI 1.3 login requests so they can be matched with authentication requests from LTI tool providers
+
+* `client_id` - {{% badge %}}LTI 1.3{{% /badge %}} client ID provided to the LTI provider
+* `login_hint` - {{% badge %}}Auto{{% /badge %}} autogenerated login hint
+* `data` - {{% badge %}}LTI 1.3{{% /badge %}} JSON blob of data about the LTI 1.3 launch
+* `createdAt` - {{% badge %}}Auto{{% /badge %}} creation timestamp
+* `updatedAt` - {{% badge %}}Auto{{% /badge %}} last updated timestamp
+
+### lti_provider_registrations
+
+This table keeps track of LTI 1.3 Dynamic Registration tokens provided to LTI 1.3 Tool Providers. These are matched with incoming registration requests to validate them and prevent unwanted registrations. 
+
+* `token` - {{% badge %}}Auto{{% /badge %}} autogenerated token
+* `url` - {{% badge %}}LTI 1.3{{% /badge %}} dynamic registration URL that the token is sent to
+* `createdAt` - {{% badge %}}Auto{{% /badge %}} creation timestamp
+* `updatedAt` - {{% badge %}}Auto{{% /badge %}} last updated timestamp
